Refactor logic.py
