+++ src/cryptoadvance/specter/specter_error.py
@@ -1,0 +1,3 @@
+class SpecterError(Exception):
+    ''' A SpecterError contains meaningfull messages which can be passed directly to the user '''
+    pass
