+++ src/cryptoadvance/specter/templates/includes/sidebar/sidebar.jinja
@@ -19,13 +19,13 @@
                 </small>
             </div>
         </a>
-        {% if specter.devices and specter.chain %}
+        {% if (specter.device_manager.devices | length) != 0 and specter.chain %}
             <div class="separator">
                 Wallets
             </div>
             {% from 'includes/sidebar/components/sidebar_wallet_list_item.jinja' import sidebar_wallet_list_item %}
-            {% for wallet in specter.wallets %}
-                {{ sidebar_wallet_list_item(wallet, wallet_alias, specter.info.chain) }}
+            {% for wallet in specter.wallet_manager.wallets %}
+                {{ sidebar_wallet_list_item(specter.wallet_manager.wallets[wallet], wallet_alias, specter.info.chain) }}
             {% endfor %}
             {{ sidebar_btn('new_wallet', 'Add new wallet') }}
         {% endif %}
@@ -34,8 +34,8 @@
             Devices
         </div>
         {% from 'includes/sidebar/components/sidebar_device_list_item.jinja' import sidebar_device_list_item %}
-        {% for device in specter.devices %}
-                {{ sidebar_device_list_item(device, device_alias) }}
+        {% for device_name in specter.device_manager.devices %}
+                {{ sidebar_device_list_item(specter.device_manager.devices[device_name], device_alias) }}
         {% endfor %}
         {{ sidebar_btn('new_device', 'Add new device') }}
         <div class="footer">
