+++ src/cryptoadvance/specter/templates/wallet/history/addresses/components/account_tx_table.jinja
@@ -14,7 +14,7 @@
             If `list_type` is 'utxo', filter out also the spent transactions of the `account` #}
         {% for tx in wallet.transactions if
             (tx["address"] == account or (wallet.getlabel(tx["address"]) == account and not group_by == 'address')) and
-            (list_type == 'history' or (wallet.istxspent(tx["txid"]) and tx["category"] != "send"))
+            (list_type == 'history' or (wallet.is_tx_spent(tx["txid"]) and tx["category"] != "send"))
         %}
             {# If the destination address is known for the transction, display the destination address, otherwise display the `account` address #}
             {% set tx_address = (tx["destination"]["address"] if ("destination" in tx and tx["destination"] != None) else tx["address"]) %}
